+++ code/packages/sdk/src/utils.ts
@@ -162,3 +162,47 @@ export function stringToColor(str: string) {
   }
   return color;
 }
+
+export function getBoundingBox(
+  stage: Konva.Stage,
+  nodes: Konva.Node[]
+): {
+  x: number;
+  y: number;
+  width: number;
+  height: number;
+} {
+  if (nodes.length === 0) {
+    return { x: 0, y: 0, width: 0, height: 0 };
+  }
+
+  let minX = Infinity;
+  let minY = Infinity;
+  let maxX = -Infinity;
+  let maxY = -Infinity;
+
+  for (const node of nodes) {
+    let realNode: Konva.Node | undefined = node;
+    if (realNode.getAttrs().containerId) {
+      realNode = stage.findOne(`#${realNode.getAttrs().containerId}`);
+    }
+
+    if (!realNode) {
+      continue;
+    }
+
+    const box = node.getRealClientRect({ skipTransform: false });
+
+    minX = Math.min(minX, box.x);
+    minY = Math.min(minY, box.y);
+    maxX = Math.max(maxX, box.x + box.width);
+    maxY = Math.max(maxY, box.y + box.height);
+  }
+
+  return {
+    x: minX,
+    y: minY,
+    width: maxX - minX,
+    height: maxY - minY,
+  };
+}